+++ src/main/java/com/matera/rest/car/CarProjection.java
@@ -12,6 +12,6 @@ public interface CarProjection {
 
     String getCarCountry();
 
-    Integer getCarOwnerId();
+    Long getCarOwnerId();
 
 }
